Provide framework for simplified s2i growth
- Added flag for modeling simplified growth within the
S2I Kinetics Butler-Volmer conditions. Currently, we only model
either this type of condition or the S2IKineticsGrowth condition,
but not in combination with each other.
- Specifically we initialize a separate dofset for this
simplified growth once such a condition is provided. We track the
growth variable (at time t_n and t_{n+1}) over the entire domain.
- The vtk output of this dofset is also implemented.
- What's currently missing for Elch:
	The growth has to be integrated in time. Currently,
	the update method of Scatra::ScatraTimIntImpl contains
	some dummy computation to see whether the vtk output
	runs. Once this integration is implemented, the
	dummy computation has to be removed.